Applications: nrf5340_audio: Print refactor
Moved functions out of server_store

part of OCT-3559
