+++ applications/nrf5340_audio/src/bluetooth/bt_stream/CMakeLists.txt
@@ -25,6 +25,7 @@ endif()
 if(CONFIG_BT_BAP_UNICAST_CLIENT)
   target_sources(app PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/unicast/unicast_client.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/unicast/unicast_client_internal.c
     ${CMAKE_CURRENT_SOURCE_DIR}/unicast/server_store.c
   )
 endif()
