+++ tests/nrf5340_audio/unicast_client_internal/prj.conf
@@ -1,0 +1,6 @@
+CONFIG_ZTEST=y
+CONFIG_ASSERT=y
+
+CONFIG_DEBUG=y
+CONFIG_LOG=y
+CONFIG_UTF8=y
