+++ tests/nrf5340_audio/unicast_client_internal/testcase.yaml
@@ -1,0 +1,12 @@
+tests:
+  nrf5340_audio.unicast_client_internal:
+    sysbuild: true
+    platform_allow:
+      - native_sim
+    integration_platforms:
+      - native_sim
+    tags:
+      - unicast_client_internal
+      - sysbuild
+      - ci_tests_nrf5340_audio
+    timeout: 20
